Update SubjectFolder to use filtering and loading pipelines.
